requirements.txt
Flask==3.1.2
requests==2.32.5
pandas==2.3.3
numpy==1.26.4
pymongo==4.9.2
mongomock==4.1.2
streamlit==1.50.0
plotly==6.0.0
python-dotenv==1.0.1
